requirements.txt
celery==5.1.2
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djongo==1.3.4
Pillow==8.2.0
psycopg2-binary>=2.8
redis==3.5.3
requests==2.26.0
